requirements.txt
openai
langchain
langchain-community
faiss-cpu
numpy
fastapi
uvicorn
python-dotenv
